Changed directory structure to permit testing and finalized the log-in page
