+++ src/main/webapp/index.html
@@ -1,0 +1,6 @@
+<!DOCTYPE html>
+<html lang="en">
+    <body>
+        <p><a href="/auth">Click Here</a></p>
+    </body>
+</html>
